requirements.txt
# متطلبات البوت الأساسية لـ Vercel
python-telegram-bot==20.7

# قاعدة البيانات
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# معالجة النصوص والوقت
emoji==2.8.0
python-dateutil==2.8.2
pytz==2023.3

# الشبكة والأمان
aiohttp==3.9.1
cryptography==41.0.7
requests==2.31.0

# معالجة البيانات
pandas==2.1.4
numpy==1.25.2

# السجلات والمراقبة
loguru==0.7.2

# البيئة والإعدادات
python-dotenv==1.0.0
pydantic==2.5.2

# اختبارات
pytest==7.4.3
pytest-asyncio==0.21.1

# إنتاج
gunicorn==21.2.0
uvicorn==0.24.0

# أدوات إضافية
pillow==10.1.0
qrcode==7.4.2
# متطلبات التطوير - Development Requirements

# الاختبارات - Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0

# تحليل الكود - Code Analysis
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3

# التوثيق - Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# أدوات التطوير - Development Tools
pre-commit==3.6.0
tox==4.11.4
bandit==1.7.5

# مراقبة الأداء - Performance Monitoring
memory-profiler==0.61.0
line-profiler==4.1.1

# أدوات إضافية - Additional Tools
ipython==8.17.2
jupyter==1.0.0
